Kill LP64 warnings. (from ross@netbsd.org)

+++ xc/programs/xterm/scrollbar.c
@@ -466,7 +466,7 @@ static void ScrollTextUpDownBy(scrollbarWidget, client_data, call_data)
 	XtPointer client_data GCC_UNUSED;
 	XtPointer call_data;
 {
-	int pixels = (int) call_data;
+	int pixels = (int)(long) call_data;
 
 	register TScreen *screen = &term->screen;
 	register int rowOnScreen, newTopLine;
@@ -556,7 +556,7 @@ void HandleScrollForward (gw, event, params, nparams)
     if (IsXtermWidget(gw)) {
     	register TScreen *screen = &((XtermWidget)gw)->screen;
 	ScrollTextUpDownBy (gw, (XtPointer) 0,
-			(XtPointer)(params_to_pixels (screen, params, *nparams)));
+	       (XtPointer)(long)(params_to_pixels (screen, params, *nparams)));
     }
     return;
 }
@@ -572,7 +572,7 @@ void HandleScrollBack (gw, event, params, nparams)
     if (IsXtermWidget(gw)) {
     	register TScreen *screen = &((XtermWidget)gw)->screen;
 	ScrollTextUpDownBy (gw, (XtPointer) 0,
-			(XtPointer)(-params_to_pixels (screen, params, *nparams)));
+	      (XtPointer)(long)(-params_to_pixels (screen, params, *nparams)));
     }
     return;
 }
